+++ ws-server/src/server.rs
@@ -42,7 +42,7 @@ use futures_util::TryStreamExt;
 use http::header::{HOST, ORIGIN};
 use http::{HeaderMap, HeaderValue};
 use jsonrpsee_core::id_providers::RandomIntegerIdProvider;
-use jsonrpsee_core::middleware::{self, WsMiddleware as Middleware};
+use jsonrpsee_core::logger::{self, WsLogger as Logger};
 use jsonrpsee_core::server::access_control::AccessControl;
 use jsonrpsee_core::server::helpers::{
 	prepare_error, BatchResponse, BatchResponseBuilder, BoundedSubscriptions, MethodResponse, MethodSink,
@@ -68,16 +68,16 @@ use tracing_futures::Instrument;
 const MAX_CONNECTIONS: u64 = 100;
 
 /// A WebSocket JSON RPC server.
-pub struct Server<M> {
+pub struct Server<L> {
 	listener: TcpListener,
 	cfg: Settings,
 	stop_monitor: StopMonitor,
 	resources: Resources,
-	middleware: M,
+	logger: L,
 	id_provider: Arc<dyn IdProvider>,
 }
 
-impl<M> std::fmt::Debug for Server<M> {
+impl<L> std::fmt::Debug for Server<L> {
 	fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
 		f.debug_struct("Server")
 			.field("listener", &self.listener)
@@ -89,7 +89,7 @@ impl<M> std::fmt::Debug for Server<M> {
 	}
 }
 
-impl<M: Middleware> Server<M> {
+impl<L: Logger> Server<L> {
 	/// Returns socket address to which the server is bound.
 	pub fn local_addr(&self) -> Result<SocketAddr, Error> {
 		self.listener.local_addr().map_err(Into::into)
@@ -116,7 +116,7 @@ impl<M: Middleware> Server<M> {
 	async fn start_inner(self, methods: Methods) {
 		let stop_monitor = self.stop_monitor;
 		let resources = self.resources;
-		let middleware = self.middleware;
+		let logger = self.logger;
 
 		let mut id = 0;
 		let mut connections = FutureDriver::default();
@@ -148,7 +148,7 @@ impl<M: Middleware> Server<M> {
 							resources: &resources,
 							cfg,
 							stop_monitor: &stop_monitor,
-							middleware: middleware.clone(),
+							logger: logger.clone(),
 							id_provider,
 						},
 					)));
@@ -218,7 +218,7 @@ where
 	}
 }
 
-enum HandshakeResponse<'a, M> {
+enum HandshakeResponse<'a, L> {
 	Reject {
 		status_code: u16,
 	},
@@ -228,12 +228,12 @@ enum HandshakeResponse<'a, M> {
 		resources: &'a Resources,
 		cfg: &'a Settings,
 		stop_monitor: &'a StopMonitor,
-		middleware: M,
+		logger: L,
 		id_provider: Arc<dyn IdProvider>,
 	},
 }
 
-async fn handshake<M: Middleware>(socket: tokio::net::TcpStream, mode: HandshakeResponse<'_, M>) -> Result<(), Error> {
+async fn handshake<L: Logger>(socket: tokio::net::TcpStream, mode: HandshakeResponse<'_, L>) -> Result<(), Error> {
 	let remote_addr = socket.peer_addr()?;
 
 	// For each incoming background_task we perform a handshake.
@@ -252,14 +252,14 @@ async fn handshake<M: Middleware>(socket: tokio::net::TcpStream, mode: Handshake
 
 			Ok(())
 		}
-		HandshakeResponse::Accept { conn_id, methods, resources, cfg, stop_monitor, middleware, id_provider } => {
+		HandshakeResponse::Accept { conn_id, methods, resources, cfg, stop_monitor, logger, id_provider } => {
 			tracing::debug!("Accepting new connection: {}", conn_id);
 
 			let key_and_headers = get_key_and_headers(&mut server, cfg).await;
 
 			match key_and_headers {
 				Ok((key, headers)) => {
-					middleware.on_connect(remote_addr, &headers);
+					logger.on_connect(remote_addr, &headers);
 					let accept = Response::Accept { key, protocol: None };
 					server.send_response(&accept).await?;
 				}
@@ -283,7 +283,7 @@ async fn handshake<M: Middleware>(socket: tokio::net::TcpStream, mode: Handshake
 				batch_requests_supported: cfg.batch_requests_supported,
 				bounded_subscriptions: BoundedSubscriptions::new(cfg.max_subscriptions_per_connection),
 				stop_server: stop_monitor.clone(),
-				middleware,
+				logger: logger,
 				id_provider,
 				ping_interval: cfg.ping_interval,
 				remote_addr,
@@ -298,7 +298,7 @@ async fn handshake<M: Middleware>(socket: tokio::net::TcpStream, mode: Handshake
 	}
 }
 
-struct BackgroundTask<'a, M> {
+struct BackgroundTask<'a, L> {
 	server: SokettoServer<'a, BufReader<BufWriter<Compat<tokio::net::TcpStream>>>>,
 	conn_id: ConnectionId,
 	methods: Methods,
@@ -309,13 +309,13 @@ struct BackgroundTask<'a, M> {
 	batch_requests_supported: bool,
 	bounded_subscriptions: BoundedSubscriptions,
 	stop_server: StopMonitor,
-	middleware: M,
+	logger: L,
 	id_provider: Arc<dyn IdProvider>,
 	ping_interval: Duration,
 	remote_addr: SocketAddr,
 }
 
-async fn background_task<M: Middleware>(input: BackgroundTask<'_, M>) -> Result<(), Error> {
+async fn background_task<L: Logger>(input: BackgroundTask<'_, L>) -> Result<(), Error> {
 	let BackgroundTask {
 		server,
 		conn_id,
@@ -327,7 +327,7 @@ async fn background_task<M: Middleware>(input: BackgroundTask<'_, M>) -> Result<
 		batch_requests_supported,
 		bounded_subscriptions,
 		stop_server,
-		middleware,
+		logger,
 		id_provider,
 		ping_interval,
 		remote_addr,
@@ -391,7 +391,7 @@ async fn background_task<M: Middleware>(input: BackgroundTask<'_, M>) -> Result<
 	// Buffer for incoming data.
 	let mut data = Vec::with_capacity(100);
 	let mut method_executors = FutureDriver::default();
-	let middleware = &middleware;
+	let logger = &logger;
 
 	let result = loop {
 		data.clear();
@@ -442,7 +442,7 @@ async fn background_task<M: Middleware>(input: BackgroundTask<'_, M>) -> Result<
 			};
 		};
 
-		let request_start = middleware.on_request();
+		let request_start = logger.on_request();
 
 		let first_non_whitespace = data.iter().find(|byte| !byte.is_ascii_whitespace());
 		match first_non_whitespace {
@@ -464,16 +464,16 @@ async fn background_task<M: Middleware>(input: BackgroundTask<'_, M>) -> Result<
 						bounded_subscriptions,
 						sink: &sink,
 						id_provider: &*id_provider,
-						middleware,
+						logger: logger,
 						request_start,
 					};
 
 					match process_single_request(data, call).await {
-						MethodResult::JustMiddleware(r) => {
-							middleware.on_response(&r.result, request_start);
+						MethodResult::JustLogger(r) => {
+							logger.on_response(&r.result, request_start);
 						}
-						MethodResult::SendAndMiddleware(r) => {
-							middleware.on_response(&r.result, request_start);
+						MethodResult::SendAndLogger(r) => {
+							logger.on_response(&r.result, request_start);
 							let _ = sink.send_raw(r.result);
 						}
 					};
@@ -487,7 +487,7 @@ async fn background_task<M: Middleware>(input: BackgroundTask<'_, M>) -> Result<
 					Id::Null,
 					ErrorObject::borrowed(BATCHES_NOT_SUPPORTED_CODE, &BATCHES_NOT_SUPPORTED_MSG, None),
 				);
-				middleware.on_response(&response.result, request_start);
+				logger.on_response(&response.result, request_start);
 				let _ = sink.send_raw(response.result);
 			}
 			Some(b'[') => {
@@ -511,14 +511,14 @@ async fn background_task<M: Middleware>(input: BackgroundTask<'_, M>) -> Result<
 							bounded_subscriptions,
 							sink: &sink,
 							id_provider: &*id_provider,
-							middleware,
+							logger: logger,
 							request_start,
 						},
 					})
 					.await;
 
 					tx_log_from_str(&response.result, max_log_length);
-					middleware.on_response(&response.result, request_start);
+					logger.on_response(&response.result, request_start);
 					let _ = sink.send_raw(response.result);
 				};
 
@@ -530,7 +530,7 @@ async fn background_task<M: Middleware>(input: BackgroundTask<'_, M>) -> Result<
 		}
 	};
 
-	middleware.on_disconnect(remote_addr);
+	logger.on_disconnect(remote_addr);
 
 	// Drive all running methods to completion.
 	// **NOTE** Do not return early in this function. This `await` needs to run to guarantee
@@ -583,10 +583,10 @@ impl Default for Settings {
 
 /// Builder to configure and create a JSON-RPC Websocket server
 #[derive(Debug)]
-pub struct Builder<M = ()> {
+pub struct Builder<L = ()> {
 	settings: Settings,
 	resources: Resources,
-	middleware: M,
+	logger: L,
 	id_provider: Arc<dyn IdProvider>,
 }
 
@@ -595,7 +595,7 @@ impl Default for Builder {
 		Builder {
 			settings: Settings::default(),
 			resources: Resources::default(),
-			middleware: (),
+			logger: (),
 			id_provider: Arc::new(RandomIntegerIdProvider),
 		}
 	}
@@ -608,7 +608,7 @@ impl Builder {
 	}
 }
 
-impl<M> Builder<M> {
+impl<L> Builder<L> {
 	/// Set the maximum size of a request body in bytes. Default is 10 MiB.
 	pub fn max_request_body_size(mut self, size: u32) -> Self {
 		self.settings.max_request_body_size = size;
@@ -650,22 +650,22 @@ impl<M> Builder<M> {
 		Ok(self)
 	}
 
-	/// Add a middleware to the builder [`Middleware`](../jsonrpsee_core/middleware/trait.Middleware.html).
+	/// Add a logger to the builder [`Logger`](../jsonrpsee_core/logger/trait.Logger.html).
 	///
 	/// ```
 	/// use std::{time::Instant, net::SocketAddr};
 	///
-	/// use jsonrpsee_core::middleware::{WsMiddleware, Headers, MethodKind, Params};
+	/// use jsonrpsee_core::logger::{WsLogger, Headers, MethodKind, Params};
 	/// use jsonrpsee_ws_server::WsServerBuilder;
 	///
 	/// #[derive(Clone)]
-	/// struct MyMiddleware;
+	/// struct MyLogger;
 	///
-	/// impl WsMiddleware for MyMiddleware {
+	/// impl WsLogger for MyLogger {
 	///     type Instant = Instant;
 	///
 	///     fn on_connect(&self, remote_addr: SocketAddr, headers: &Headers) {
-	///          println!("[MyMiddleware::on_call] remote_addr: {}, headers: {:?}", remote_addr, headers);
+	///          println!("[MyLogger::on_call] remote_addr: {}, headers: {:?}", remote_addr, headers);
 	///     }
 	///
 	///     fn on_request(&self) -> Self::Instant {
@@ -673,26 +673,26 @@ impl<M> Builder<M> {
 	///     }
 	///
 	///     fn on_call(&self, method_name: &str, params: Params, kind: MethodKind) {
-	///          println!("[MyMiddleware::on_call] method: '{}' params: {:?}, kind: {:?}", method_name, params, kind);
+	///          println!("[MyLogger::on_call] method: '{}' params: {:?}, kind: {:?}", method_name, params, kind);
 	///     }
 	///
 	///     fn on_result(&self, method_name: &str, success: bool, started_at: Self::Instant) {
-	///          println!("[MyMiddleware::on_result] '{}', worked? {}, time elapsed {:?}", method_name, success, started_at.elapsed());
+	///          println!("[MyLogger::on_result] '{}', worked? {}, time elapsed {:?}", method_name, success, started_at.elapsed());
 	///     }
 	///
 	///     fn on_response(&self, result: &str, started_at: Self::Instant) {
-	///          println!("[MyMiddleware::on_response] result: {}, time elapsed {:?}", result, started_at.elapsed());
+	///          println!("[MyLogger::on_response] result: {}, time elapsed {:?}", result, started_at.elapsed());
 	///     }
 	///
 	///     fn on_disconnect(&self, remote_addr: SocketAddr) {
-	///          println!("[MyMiddleware::on_disconnect] remote_addr: {}", remote_addr);
+	///          println!("[MyLogger::on_disconnect] remote_addr: {}", remote_addr);
 	///     }
 	/// }
 	///
-	/// let builder = WsServerBuilder::new().set_middleware(MyMiddleware);
+	/// let builder = WsServerBuilder::new().set_logger(MyLogger);
 	/// ```
-	pub fn set_middleware<T: Middleware>(self, middleware: T) -> Builder<T> {
-		Builder { settings: self.settings, resources: self.resources, middleware, id_provider: self.id_provider }
+	pub fn set_logger<T: Logger>(self, logger: T) -> Builder<T> {
+		Builder { settings: self.settings, resources: self.resources, logger, id_provider: self.id_provider }
 	}
 
 	/// Configure a custom [`tokio::runtime::Handle`] to run the server on.
@@ -771,7 +771,7 @@ impl<M> Builder<M> {
 	/// }
 	/// ```
 	///
-	pub async fn build(self, addrs: impl ToSocketAddrs) -> Result<Server<M>, Error> {
+	pub async fn build(self, addrs: impl ToSocketAddrs) -> Result<Server<L>, Error> {
 		let listener = TcpListener::bind(addrs).await?;
 		let stop_monitor = StopMonitor::new();
 		let resources = self.resources;
@@ -780,7 +780,7 @@ impl<M> Builder<M> {
 			cfg: self.settings,
 			stop_monitor,
 			resources,
-			middleware: self.middleware,
+			logger: self.logger,
 			id_provider: self.id_provider,
 		})
 	}
@@ -805,43 +805,43 @@ async fn send_ws_ping(sender: &mut Sender<BufReader<BufWriter<Compat<TcpStream>>
 }
 
 #[derive(Debug, Clone)]
-struct Batch<'a, M: Middleware> {
+struct Batch<'a, L: Logger> {
 	data: Vec<u8>,
-	call: CallData<'a, M>,
+	call: CallData<'a, L>,
 }
 
 #[derive(Debug, Clone)]
-struct CallData<'a, M: Middleware> {
+struct CallData<'a, L: Logger> {
 	conn_id: usize,
 	bounded_subscriptions: BoundedSubscriptions,
 	id_provider: &'a dyn IdProvider,
-	middleware: &'a M,
+	logger: &'a L,
 	methods: &'a Methods,
 	max_response_body_size: u32,
 	max_log_length: u32,
 	resources: &'a Resources,
 	sink: &'a MethodSink,
-	request_start: M::Instant,
+	request_start: L::Instant,
 }
 
 #[derive(Debug, Clone)]
-struct Call<'a, M: Middleware> {
+struct Call<'a, L: Logger> {
 	params: Params<'a>,
 	name: &'a str,
-	call: CallData<'a, M>,
+	call: CallData<'a, L>,
 	id: Id<'a>,
 }
 
 enum MethodResult {
-	JustMiddleware(MethodResponse),
-	SendAndMiddleware(MethodResponse),
+	JustLogger(MethodResponse),
+	SendAndLogger(MethodResponse),
 }
 
 impl MethodResult {
 	fn as_inner(&self) -> &MethodResponse {
 		match &self {
-			Self::JustMiddleware(r) => r,
-			Self::SendAndMiddleware(r) => r,
+			Self::JustLogger(r) => r,
+			Self::SendAndLogger(r) => r,
 		}
 	}
 }
@@ -849,9 +849,9 @@ impl MethodResult {
 // Batch responses must be sent back as a single message so we read the results from each
 // request in the batch and read the results off of a new channel, `rx_batch`, and then send the
 // complete batch response back to the client over `tx`.
-async fn process_batch_request<M>(b: Batch<'_, M>) -> BatchResponse
+async fn process_batch_request<L>(b: Batch<'_, L>) -> BatchResponse
 where
-	M: Middleware,
+	L: Logger,
 {
 	let Batch { data, call } = b;
 
@@ -861,27 +861,28 @@ where
 			let batch_stream = futures_util::stream::iter(batch);
 
 			let trace = RpcTracing::batch();
-			let _enter = trace.span().enter();
-			let max_response_size = call.max_response_body_size;
 
-			let batch_response = batch_stream
-				.try_fold(
-					BatchResponseBuilder::new_with_limit(max_response_size as usize),
-					|batch_response, (req, call)| async move {
-						let params = Params::new(req.params.map(|params| params.get()));
+			return async {
+				let max_response_size = call.max_response_body_size;
 
-						let response =
-							execute_call(Call { name: &req.method, params, id: req.id, call }).in_current_span().await;
+				let batch_response = batch_stream
+					.try_fold(
+						BatchResponseBuilder::new_with_limit(max_response_size as usize),
+						|batch_response, (req, call)| async move {
+							let params = Params::new(req.params.map(|params| params.get()));
+							let response = execute_call(Call { name: &req.method, params, id: req.id, call }).await;
+							batch_response.append(response.as_inner())
+						},
+					)
+					.await;
 
-						batch_response.append(response.as_inner())
-					},
-				)
-				.await;
-
-			return match batch_response {
-				Ok(batch) => batch.finish(),
-				Err(batch_err) => batch_err,
-			};
+				match batch_response {
+					Ok(batch) => batch.finish(),
+					Err(batch_err) => batch_err,
+				}
+			}
+			.instrument(trace.into_span())
+			.await;
 		} else {
 			BatchResponse::error(Id::Null, ErrorObject::from(ErrorCode::InvalidRequest))
 		};
@@ -891,21 +892,24 @@ where
 	BatchResponse::error(id, ErrorObject::from(code))
 }
 
-async fn process_single_request<M: Middleware>(data: Vec<u8>, call: CallData<'_, M>) -> MethodResult {
+async fn process_single_request<L: Logger>(data: Vec<u8>, call: CallData<'_, L>) -> MethodResult {
 	if let Ok(req) = serde_json::from_slice::<Request>(&data) {
 		let trace = RpcTracing::method_call(&req.method);
-		let _enter = trace.span().enter();
 
-		rx_log_from_json(&req, call.max_log_length);
+		async {
+			rx_log_from_json(&req, call.max_log_length);
 
-		let params = Params::new(req.params.map(|params| params.get()));
-		let name = &req.method;
-		let id = req.id;
+			let params = Params::new(req.params.map(|params| params.get()));
+			let name = &req.method;
+			let id = req.id;
 
-		execute_call(Call { name, params, id, call }).in_current_span().await
+			execute_call(Call { name, params, id, call }).await
+		}
+		.instrument(trace.into_span())
+		.await
 	} else {
 		let (id, code) = prepare_error(&data);
-		MethodResult::SendAndMiddleware(MethodResponse::error(id, ErrorObject::from(code)))
+		MethodResult::SendAndLogger(MethodResponse::error(id, ErrorObject::from(code)))
 	}
 }
 
@@ -914,12 +918,12 @@ async fn process_single_request<M: Middleware>(data: Vec<u8>, call: CallData<'_,
 ///
 /// Returns `(MethodResponse, None)` on every call that isn't a subscription
 /// Otherwise `(MethodResponse, Some(PendingSubscriptionCallTx)`.
-async fn execute_call<M: Middleware>(c: Call<'_, M>) -> MethodResult {
+async fn execute_call<L: Logger>(c: Call<'_, L>) -> MethodResult {
 	let Call { name, id, params, call } = c;
 	let CallData {
 		resources,
 		methods,
-		middleware,
+		logger,
 		max_response_body_size,
 		max_log_length,
 		conn_id,
@@ -931,29 +935,29 @@ async fn execute_call<M: Middleware>(c: Call<'_, M>) -> MethodResult {
 
 	let response = match methods.method_with_name(name) {
 		None => {
-			middleware.on_call(name, params.clone(), middleware::MethodKind::Unknown);
+			logger.on_call(name, params.clone(), logger::MethodKind::Unknown);
 			let response = MethodResponse::error(id, ErrorObject::from(ErrorCode::MethodNotFound));
-			MethodResult::SendAndMiddleware(response)
+			MethodResult::SendAndLogger(response)
 		}
 		Some((name, method)) => match &method.inner() {
 			MethodKind::Sync(callback) => {
-				middleware.on_call(name, params.clone(), middleware::MethodKind::MethodCall);
+				logger.on_call(name, params.clone(), logger::MethodKind::MethodCall);
 
 				match method.claim(name, resources) {
 					Ok(guard) => {
 						let r = (callback)(id, params, max_response_body_size as usize);
 						drop(guard);
-						MethodResult::SendAndMiddleware(r)
+						MethodResult::SendAndLogger(r)
 					}
 					Err(err) => {
 						tracing::error!("[Methods::execute_with_resources] failed to lock resources: {:?}", err);
 						let response = MethodResponse::error(id, ErrorObject::from(ErrorCode::ServerIsBusy));
-						MethodResult::SendAndMiddleware(response)
+						MethodResult::SendAndLogger(response)
 					}
 				}
 			}
 			MethodKind::Async(callback) => {
-				middleware.on_call(name, params.clone(), middleware::MethodKind::MethodCall);
+				logger.on_call(name, params.clone(), logger::MethodKind::MethodCall);
 
 				match method.claim(name, resources) {
 					Ok(guard) => {
@@ -962,43 +966,43 @@ async fn execute_call<M: Middleware>(c: Call<'_, M>) -> MethodResult {
 
 						let response =
 							(callback)(id, params, conn_id, max_response_body_size as usize, Some(guard)).await;
-						MethodResult::SendAndMiddleware(response)
+						MethodResult::SendAndLogger(response)
 					}
 					Err(err) => {
 						tracing::error!("[Methods::execute_with_resources] failed to lock resources: {:?}", err);
 						let response = MethodResponse::error(id, ErrorObject::from(ErrorCode::ServerIsBusy));
-						MethodResult::SendAndMiddleware(response)
+						MethodResult::SendAndLogger(response)
 					}
 				}
 			}
 			MethodKind::Subscription(callback) => {
-				middleware.on_call(name, params.clone(), middleware::MethodKind::Subscription);
+				logger.on_call(name, params.clone(), logger::MethodKind::Subscription);
 
 				match method.claim(name, resources) {
 					Ok(guard) => {
 						if let Some(cn) = bounded_subscriptions.acquire() {
 							let conn_state = ConnState { conn_id, close_notify: cn, id_provider };
 							let response = callback(id.clone(), params, sink.clone(), conn_state, Some(guard)).await;
-							MethodResult::JustMiddleware(response)
+							MethodResult::JustLogger(response)
 						} else {
 							let response =
 								MethodResponse::error(id, reject_too_many_subscriptions(bounded_subscriptions.max()));
-							MethodResult::SendAndMiddleware(response)
+							MethodResult::SendAndLogger(response)
 						}
 					}
 					Err(err) => {
 						tracing::error!("[Methods::execute_with_resources] failed to lock resources: {:?}", err);
 						let response = MethodResponse::error(id, ErrorObject::from(ErrorCode::ServerIsBusy));
-						MethodResult::SendAndMiddleware(response)
+						MethodResult::SendAndLogger(response)
 					}
 				}
 			}
 			MethodKind::Unsubscription(callback) => {
-				middleware.on_call(name, params.clone(), middleware::MethodKind::Unsubscription);
+				logger.on_call(name, params.clone(), logger::MethodKind::Unsubscription);
 
 				// Don't adhere to any resource or subscription limits; always let unsubscribing happen!
 				let result = callback(id, params, conn_id, max_response_body_size as usize);
-				MethodResult::SendAndMiddleware(result)
+				MethodResult::SendAndLogger(result)
 			}
 		},
 	};
@@ -1006,7 +1010,7 @@ async fn execute_call<M: Middleware>(c: Call<'_, M>) -> MethodResult {
 	let r = response.as_inner();
 
 	rx_log_from_str(&r.result, max_log_length);
-	middleware.on_result(name, r.success, request_start);
+	logger.on_result(name, r.success, request_start);
 	response
 }